feat(): integrated mongodb, add jwt authorization; rewrite controllers route

+++ Models/PageInfo.cs
@@ -3,11 +3,11 @@ namespace csi5112group1project_service.Models;
 public class PageInfo<T>
 {
   public T rows { get; set; }
-  public int totalRows { get; set; }
+  public long totalRows { get; set; }
 
   public PageInfo(
     T rows,
-    int totalRows
+    long totalRows
   )
   {
     this.rows = rows;
